third_party/WebKit/LayoutTests/fast/canvas/OffscreenCanvas-commit-copyImage.html

<html>
<body><canvas id="myCanvas" width="256" height="256"></canvas>
  <div id="log"></div>
</body>
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<script>

var t = async_test("CopyImageOnTransferredCanvas");

function main() {
  assert_true(!!window.testRunner, "This test requires window.testRunner.");

  if (window.testRunner) {
    initTest();
  } else {
    t.done();
  }
}

function initTest() {
  var canv = document.getElementById("myCanvas");
  var off = canv.transferControlToOffscreen();
  var ctx = off.getContext("2d");

  ctx.fillStyle = "#f00";
  ctx.fillRect(0, 0, 128, 128);
  ctx.fillStyle = "#0f0";
  ctx.fillRect(128, 0, 128, 128);
  ctx.fillStyle = "#00f";
  ctx.fillRect(0, 128, 128, 128);
  ctx.fillStyle = "#ff0";
  ctx.fillRect(128, 128, 128, 128);

  requestAnimationFrame(runTest);
}

function copyImage() {
  testRunner.copyImageAtAndCapturePixelsAsyncThen(100, 100, completionCallback);
}

function runTest() {
    // We need setTimeout as a frame barrier for the image frame to be
    // propagated to the placeholder canvas.
    requestAnimationFrame(() => {
      requestAnimationFrame(() => {
        requestAnimationFrame(copyImage);
      });
    });
}

function completionCallback(width, height, snapshot) {
    t.step(function() {
      assert_equals(width, 256, "The copied image has a width of 256.");
      assert_equals(height, 256, "The copied image has a height of 256.");

      var data = new Uint8Array(snapshot);

      var pixelLeftTop = data.subarray(0, 4);
      assert_array_equals(pixelLeftTop, [255, 0, 0, 255], "The copied image's top left is red");

      var pixelRightTop = data.subarray(4 * 128, 4* 128 + 4);
      assert_array_equals(pixelRightTop, [0, 255, 0, 255], "The copied image's top right is green");

      var pixelLeftBottom = data.subarray(4 * 256 * 128, 4 * 256 * 128 + 4);
      assert_array_equals(pixelLeftBottom, [0, 0, 255, 255], "The copied image's bottom left is blue");

      var rightBottomPixelPosition = 4 * 256 * 128 + 4 * 128;
      var pixelRightBottom = data.subarray(rightBottomPixelPosition, rightBottomPixelPosition + 4);
      assert_array_equals(pixelRightBottom, [255, 255, 0, 255], "The copied image's bottom right is yellow");

    });
    t.done();
}

main();
  </script>
</html>
